==> code/5_moisture_model.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# 0. Load packages

```{r}
library(tidyverse)
library(caret)
library(xgboost)
library(ModelMetrics)
```

# 1. Import data

```{r}
df_etc <- read.csv("../data/procesed/etc_pred_copy.csv") %>% .[-1]
df_sta <- read.csv("../data/procesed/static_df.csv") %>% select(-X)
df_dyn <- read.csv("../data/procesed/dynamic_df.csv") %>% select(-X)
df_sm <- read.csv('../data/procesed/soil_aqua_sentek_6inch.csv') %>% .[-c(1,5)] %>% 
  mutate(Date = yday(as.Date(Date)))

summary(df_sm)

df.xgb <- df_sm %>% 
  left_join(
  df_etc) %>% 
  left_join(df_sta) %>% 
  left_join(df_dyn) %>% 
  drop_na() %>% 
  unite(team_id, Team, Plot_ID, remove = F)
```

# 2. Model

$${Arable~ET_c} \sim f(\text{Management}_{\text{plot}}, \text{EC}_{\text{plot}}, \text{Soil texture}_{\text{plot}}, \text{NDVI}_{\text{daily}}, \text{Weather}_{\text{daily}}, \text{Irrigation}_{\text{daily}})$$

# Model: sm 

```{r}
train_treatments <- df.xgb %>%
  group_by(team_id) %>%
  summarise() %>%
  sample_frac(0.7) %>%
  pull(team_id)

df_train <- df.xgb %>%
  filter(team_id %in% train_treatments)

df_test <- df.xgb %>%
  filter(!team_id %in% train_treatments)
```

## XgBoost

```{r}
tune_control <- caret::trainControl(
  method = "cv",
  number = 5,
  verboseIter = FALSE,
  allowParallel = TRUE)

hyperparam_grid <- expand.grid(
  nrounds = c(100, 300),
  max_depth = c(3, 5, 10),
  eta = c(0.005, 0.01, 0.1),
  gamma = 0,
  colsample_bytree = c(0.7, 1),
  min_child_weight = 1,
  subsample = c(0.7, 1))

X_train <- df_train %>%
  dplyr::select(-c(Team, team_id, Block_ID, Plot_ID, daily_avg_moist)) %>%
  as.matrix()

y_train <- df_train$daily_avg_moist

# Hyperparameter tuning using k-fold cross-validation
sm_model <- caret::train(
  x = X_train,
  y = y_train,
  trControl = tune_control,
  tuneGrid = hyperparam_grid,
  method = "xgbTree",
  verbose = FALSE,
  verbosity = 0)

best_params_sm <- sm_model$bestTune

sm_preds_final <- data.frame(team_id = integer(), Date = as.integer(), preds = numeric(), observed_sm = numeric())
sm_importance_list <- list()
unique_treatments <- unique(df.xgb$team_id)

for (i in unique_treatments) {
  
  df_train_loocv <- df.xgb %>% filter(team_id != i)
  df_test_loocv <- df.xgb %>% filter(team_id == i)
  
  X_train_loocv <- df_train_loocv %>% select(-c(Team, team_id, Block_ID, Plot_ID, daily_avg_moist)) %>% as.matrix()
  y_train_loocv <- df_train_loocv$daily_avg_moist
  
  X_test_loocv <- df_test_loocv %>% select(-c(Team, team_id, Block_ID, Plot_ID, daily_avg_moist)) %>% as.matrix()
  y_test_loocv <- df_test_loocv$daily_avg_moist

  final_sm_model <- xgboost(
    data = X_train_loocv,
    label = y_train_loocv,
    booster = "gbtree",
    objective = "reg:squarederror",
    nrounds = best_params_sm$nrounds,
    max_depth = best_params_sm$max_depth,
    colsample_bytree = best_params_sm$colsample_bytree,
    min_child_weight = best_params_sm$min_child_weight,
    subsample = best_params_sm$subsample,
    eta = best_params_sm$eta,
    gamma = best_params_sm$gamma,
    verbose = 0)
  
  temp_pred_sm <- predict(final_sm_model, X_test_loocv)
  
  preds <- data.frame(team_id = i, date = df_test_loocv$Date, preds = temp_pred_sm, observed_sm = y_test_loocv)
  
  sm_preds_final <- rbind(sm_preds_final, preds)
  
  importance <- xgb.importance(model = final_sm_model)
  importance$iteration <- i
  sm_importance_list[[as.character(i)]] <- importance
}

sm_rmse <- rmse(sm_preds_final$observed_sm, sm_preds_final$preds)
sm_mae <- mae(sm_preds_final$observed_sm, sm_preds_final$preds)
sm_r2 <- cor(sm_preds_final$observed_sm, sm_preds_final$preds)^2

list(sm_RMSE = sm_rmse, sm_MAE = sm_mae, sm_squared = sm_r2)
```

## Preds

```{r}
sm_importance_df <- do.call(rbind, sm_importance_list) %>%
  group_by(Feature) %>%
  summarise(Importance = mean(Gain, na.rm = TRUE)) %>%
  arrange(desc(Importance))

top_15_sm <- sm_importance_df %>% top_n(15, Importance)

ggplot(top_15_sm, aes(x = reorder(Feature, Importance), y = Importance, fill = Importance)) +
  geom_bar(stat = "identity", show.legend = FALSE) +
  coord_flip() +
  scale_fill_gradient(low = "grey", high = "purple") +
  labs(title = "Variable Importance for Soil Moisture Prediction (Top 15)",
       x = "Features",
       y = "Gain") +
  theme_minimal(base_size = 15) +
  theme(
    plot.title = element_text(hjust = 0.5, face = "bold", size = 20),
    axis.title.x = element_text(face = "bold", size = 15),
    axis.title.y = element_text(face = "bold", size = 15),
    axis.text = element_text(size = 12))

df_sm_preds_final <- sm_preds_final %>% na.omit()

sm_r_squared <- cor(df_sm_preds_final$observed_sm, df_sm_preds_final$preds)^2
sm_mae <- mean(abs(df_sm_preds_final$observed_sm - df_sm_preds_final$preds))
sm_rmse <- sqrt(mean((df_sm_preds_final$observed_sm - df_sm_preds_final$preds)^2))
sm_rrmse <- sm_rmse / mean(df_sm_preds_final$observed_sm) * 100
sm_kge <- 1 - sqrt((sm_r_squared - 1)^2 + (mean(df_sm_preds_final$preds) / mean(df_sm_preds_final$observed_sm) - 1)^2)

ggplot(df_sm_preds_final, aes(x = observed_sm, y = preds)) +
  geom_point(shape = 21, color = "grey", fill = "steelblue", alpha = 0.5, size = 3) +
  geom_abline(slope = 1, intercept = 0, color = "black", size = 1) + # 1:1 line
  xlim(20, 70) +
  ylim(20, 70) +
  theme_bw() +
  labs(x = "Observed Soil Moisture", y = "Predicted Soil Moisture") +
  # Add metrics as text
  annotate("text", x = 20, y = 70, 
           label = paste0(
             "R²: ", round(sm_r_squared, 3), "\n",
             "MAE: ", round(sm_mae, 3), "\n",
             "RMSE: ", round(sm_rmse, 3), "\n",
             "RRMSE: ", round(sm_rrmse, 1), "%\n",
             "KGE: ", round(sm_kge, 3)
           ), hjust = 0, vjust = 1) +
  theme(aspect.ratio = 1,
        text = element_text(size = 18))
```

## 

```{r}

full_sm_model <- xgboost(
  data = X_train,
  label = y_train,
  booster = "gbtree",
  objective = "reg:squarederror",
  nrounds = best_params_sm$nrounds,
  max_depth = best_params_sm$max_depth,
  colsample_bytree = best_params_sm$colsample_bytree,
  min_child_weight = best_params_sm$min_child_weight,
  subsample = best_params_sm$subsample,
  eta = best_params_sm$eta,
  gamma = best_params_sm$gamma,
  verbose = 0
)

df_train%>%
  dplyr::select(-c(Team, team_id, Block_ID, Plot_ID, daily_avg_moist))

df_sm_new <- df_dyn %>%
  left_join(df_sta, by = c("Team", "Plot_ID")) %>%
  drop_na() %>%
  select(-c(Team, Block_ID, Plot_ID)) %>%
  select(
    Date,
    Sand,
    Clay,
    SOM,
    EC_Shallow,
    EC_deep,
    Planting_date,
    Seeding_rate,
    N_rate,
    TEMP2MMIN,
    TEMP2MMAX,
    RELHUM2MMAX ,
    RELHUM2MMIN ,
    PRECIP ,
    SRAVG ,
    WSPD2MAVG ,
    SOILTMP5AVG ,
    Irrigation_mm,
    NDVI_pred) %>% 
  left_join(df_etc[,-c(21:23)])

df_sm_new$daily_avg_moist_pred <- predict(full_sm_model, as.matrix(df_sm_new))

df_sm_new <- df_sm_new %>% 
  left_join(df_sta)

df_sm_new %>% 
  ggplot()+
  geom_line(aes(x=Date, y=daily_avg_moist_pred, color=factor(Plot_ID)), show.legend = F, alpha = 0.5)+
  theme_bw()+
  xlab("DOY")+
  ylab("Predicted Soil Moisture (%)")+
  theme(text = element_text(size = 18),
        panel.grid = element_blank())

#df_sm_new %>% write.csv("../data/procesed/sm_pred.csv")
```
